Resolvendo alguns bugs
